create based combined field handler and move common logic into it (task #2070)
